+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - BREAKING feat(): `startImmediatePropagation` [#9099](https://github.com/fabricjs/fabric.js/pull/9099)
   **BREAKING** `setActiveObject`, `_setActiveObject` signature => `(object, e)` => `(object, options?: { e?, startImmediatePropagation? })`
+- ci(e2e): support relative imports [#9108](https://github.com/fabricjs/fabric.js/pull/9108)
 - fix(): #8344 stroke projection [#8374](https://github.com/fabricjs/fabric.js/pull/8374)
 - fix(Filters) Removing type from the options passed in the constructor [#9089](https://github.com/fabricjs/fabric.js/pull/9089)
 - feat(InteractiveObject): add `getActiveControl()` to expose `__corner` [#9102](https://github.com/fabricjs/fabric.js/pull/9102)